requirements.txt
# Project dependencies
click>=8.0.0
pyyaml>=6.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
pandas>=2.0.0

# FastAPI and API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.5.0

# WebSocket and real-time features
python-socketio>=5.10.0
python-multipart>=0.0.6

# Additional dependencies for web scraping and AWS integration
selenium>=4.15.0
boto3>=1.28.0
botocore>=1.31.0
sqlalchemy-utils>=0.41.0
sqlmodel>=0.0.10
aiohttp>=3.8.0
ratelimit>=2.2.0

# Additional dependencies for LLM integration
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
langchain-aws>=0.1.0
langchain-huggingface>=0.0.1

# Additional dependencies for embeddings
huggingface_hub>=0.24.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
gcsfs>=2023.5.0
unstructured>=0.16.5
tiktoken
bs4
typing_extensions